+++ options/Sounds.js
@@ -5,10 +5,12 @@ import browser from "webextension-polyfill";
 import Sound from '../common/sound';
 import { emptyObject } from '../common/utils';
 import { shrinkFont } from './utils';
+import throttle from 'just-throttle';
 
 class Sounds {
   constructor(el, parentStore) {
     this.store = parentStore;
+    this.instantPlay = false;
 
     this.$el   = el;
     this.$list = this.$el.querySelector('ul.list');
@@ -16,12 +18,19 @@ class Sounds {
 
     this._$selected = null;
     this._observers = {};
+    this.$togglePickPlay = document.querySelector('#instant-play-toggle input[type="checkbox"]');
     this.$addSound = this.$el.querySelector('.add_sound');
 
+    this.$dragging = null;
     this.$list.addEventListener('focus', this.onFocus.bind(this), {capture: true});
     this.$list.addEventListener('click', this.onSelect.bind(this));
     this.$list.addEventListener('keydown', this.onKey.bind(this));
-    this.$addSound.addEventListener('click', () => this.$selected = this.addSound());
+    this.$list.addEventListener('dragstart', this.onDragStart.bind(this));
+    this.$list.addEventListener('dragover', throttle(this.onDragOver.bind(this)), 900);
+    this.$list.addEventListener('drop', this.onDrop.bind(this));
+    this.$list.addEventListener('dragend', this.onDragEnd.bind(this));
+    this.$addSound.addEventListener('click', this.newSound.bind(this));
+    this.$togglePickPlay.addEventListener('change', (e) => this.instantPlay = e.target.checked);
     this.load();
   }
 
@@ -67,15 +76,26 @@ class Sounds {
 
   onFocus(e) {
     let $li = e.target.closest('.list li');
-    if ($li !== this.$addSound) {
+
+    if (!this.instantPlay && $li !== this.$addSound) {
       this.$selected = $li;
     }
   }
 
   onSelect(e) {
     let $li = e.target.closest('.list li');
+
+    if (!$li) { // click outside of sounds
+      this.$selected = null;
+      return;
+    }
+
     if ($li !== this.$addSound) {
       this.$selected = $li;
+
+      if (this.instantPlay) {
+        this.notifyObservers('testPlay');
+      }
     }
   }
 
@@ -87,10 +107,69 @@ class Sounds {
     }
   }
 
-  set $selected(v) {
-    if (this._$selected) {
-      this._$selected.classList.remove('current');
+  onDragStart(e) {
+    const $li = e.target;
+    if ($li.tagName === 'LI' && !$li.classList.contains('add_sound')) {
+      this.$dragging = $li;
     }
+  }
+
+  onDragEnd(e) {
+    this.$dragging = null;
+    this.$list.querySelectorAll('.drag-before, .drag-after').forEach((otherLi) => {
+      otherLi.classList.remove('drag-before', 'drag-after');
+    });
+  }
+
+  findDropee(el) {
+    if (!this.$dragging) return;
+
+    const tagName = el.tagName;
+    let $li = tagName === 'LI' ? el : (
+      ['LABEL', 'SPAN'].includes(tagName) ? el.closest('li[draggable]') : null
+    );
+
+    if ($li?.draggable && !$li.classList.contains('add_sound') && $li !== this.$dragging) {
+      return $li;
+    }
+  }
+
+  findDropSide(el, clientX) {
+    const boundingBox = el.getBoundingClientRect();
+    const offset = boundingBox.x + boundingBox.width / 2;
+    return clientX < offset;
+  }
+
+  onDragOver(e) {
+    e.preventDefault();
+    const $li = this.findDropee(e.target);
+    if (!$li) return;
+
+    this.$list.querySelectorAll('.drag-before, .drag-after').forEach((otherLi) => {
+      otherLi.classList.remove('drag-before', 'drag-after');
+    });
+
+    if (this.findDropSide($li, e.clientX)) {
+      $li.classList.add('drag-before');
+    } else {
+      $li.classList.add('drag-after');
+    }
+  }
+
+  onDrop(e) {
+    e.preventDefault();
+    const $li = this.findDropee(e.target);
+    if (!$li) return;
+
+    if (this.findDropSide($li, e.clientX)) {
+      this.$dragging.parentNode.insertBefore(this.$dragging, $li);
+    } else {
+      this.$dragging.parentNode.insertBefore(this.$dragging, $li.nextElementSibling);
+    }
+  }
+
+  set $selected(v) {
+    this.$list.querySelectorAll('li.current').forEach(el => el.classList.remove('current'));
     this._$selected = v;
     if (this._$selected) {
       this._$selected.classList.add('current');
@@ -131,6 +210,11 @@ class Sounds {
         el.parentNode.insertBefore(el, el.parentNode.firstElementChild);
       }
     }
+  }
+
+  newSound() {
+    this.$selected = this.addSound();
+    this.notifyObservers('new');
   }
 
   addSound(config) {